修改了Blog生成临时文件的目录

+++ Blog/filters/CacheFilter.php
@@ -1,6 +1,8 @@
 <?php
 namespace Blog\filters;
 
+use Arsenals\Core\file_exists;
+
 use Arsenals\Core\Abstracts\Filter;
 /**
  * 缓存过滤器
@@ -17,10 +19,11 @@ class CacheFilter implements Filter {
 		$path_info = $router->getPathInfo();
 		
 		// 所有文章页面内容缓存
-		if(\Arsenals\Core\str_start_with($path_info, 'article_') ){
+		if(\Arsenals\Core\str_start_with($path_info, 'article/') ){
 			//$cache_file = CACHE_PATH . md5($path_info);
            
 			$cache_file = BASE_PATH . $path_info;
+			// 如果是SAE，则声称MD5的标识
             if(defined('IS_SAE') && IS_SAE){
             	$cache_file = md5($cache_file);
             }
@@ -39,7 +42,9 @@ class CacheFilter implements Filter {
 			$filterChain->doFilter();
 			$content = ob_get_contents();
 			ob_end_clean();
-			\Arsenals\Core\file_put_contents($cache_file, $content);
+			if(\file_exists(dirname($cache_file))){
+				\Arsenals\Core\file_put_contents($cache_file, $content);
+			}
 			echo $content;
 			return true;
 		}
